[Articles] Moved article-related components to articles folder
