+++ lib/components/artist/articles/article.js
@@ -5,9 +5,9 @@ import Relay from 'react-relay';
 import React from 'react-native';
 const { ScrollView, StyleSheet, View, Text } = React;
 
-import colors from '../../../data/colors';
-import SerifText from '../text/serif';
-import ImageView from '../opaque_image_view';
+import colors from '../../../../data/colors';
+import SerifText from '../../text/serif';
+import ImageView from '../../opaque_image_view';
 
 class Article extends React.Component {
   render() {
